Updated project using Xcode 15.4
- updated to recommended project settings
- raised deployment targets to the minimum supported - iOS 12
- bumped library version to 1.9.0

In addition to that there were 2 tests failing related to type castig NSNumber to Float after decoding.
It turned out that since iOS 17 and Xcode 15, NSNumber created with Double cannot be type casted to Float, but has to be converted explicitly via the NSNumber API. This kind of make sense, because JSON does not have explicit types for Double/Float/Int etc.. and every number with decimal precision is represented as Double.